+++ lib/ws_servers/api/send_authenticated.js
@@ -58,21 +58,5 @@ module.exports = async (server, ws, opts) => {
   }])
   send(ws, ['data.api_credentials.configured', 'bitfinex'])
 
-  const results = await connectionManager.start(server, ws)
-  let status = {
-    dmsControl: false,
-    algoWorker: false,
-    bfxClient: false,
-    strategyManager: false
-  }
-
-  for (const { reason: err, value } of results) {
-    if (err) {
-      sendError(ws, `Failed to start services: ${err.message}`)
-    } else {
-      status = { ...status, ...value }
-    }
-  }
-
-  send(ws, ['info.services.status', mode, status])
+  await connectionManager.start(server, ws)
 }
